Update to 20160301.

+++ sys/dev/auvolconv.c
@@ -63,7 +63,7 @@ auvolconv_slinear16_le_fetch_to(struct audio_softc *asc,
 	FILTER_LOOP_PROLOGUE(this->src, 2, dst, 2, m) {
 		j = le16dec(s);
 		wp = (int16_t *)d;
-		le16enc(wp, (j * vol) / 255);
+		le16enc(wp, (int32_t)(j * vol) / 255);
 	} FILTER_LOOP_EPILOGUE(this->src, dst);
 
 	return 0;
@@ -91,7 +91,7 @@ auvolconv_slinear16_be_fetch_to(struct audio_softc *asc,
 	FILTER_LOOP_PROLOGUE(this->src, 2, dst, 2, m) {
 		j = be16dec(s);
 		wp = (int16_t *)d;
-		be16enc(wp, (j * vol) / 255);
+		be16enc(wp, (int32_t)(j * vol) / 255);
 	} FILTER_LOOP_EPILOGUE(this->src, dst);
 
 	return 0;
